Made ggml_interface_fp16 independent
Removed ggml includes, renamed some functions, moved some declerations to common nntr_ggml_impl headers
