+++ nntrainer/tensor/cpu_backend/arm/arm_compute_backend_fp16.cpp
@@ -330,7 +330,7 @@ size_t quantize_q8_0(const _FP16 *src, void *dst, int64_t nrow,
 
 template <> void dequantize_row_q8_0(const void *x_raw, _FP16 *y, int64_t k) {
 #ifdef ENABLE_GGML
-  __nntr_dequantize_row_q8_0(x_raw, y, k);
+  __ggml_dequantize_row_q8_0(x_raw, y, k);
 #else
   __fallback_dequantize_row_q8_0(x_raw, y, k);
 #endif
